scripts/redbook/requirements.txt
# @AI_GENERATED
playwright
python-dotenv
hypothesis
pytest
pytest-asyncio
pytest-playwright
openai
# @AI_GENERATED: end
